adds `pollReactionHandler`
