+++ src/applications/floor-sweeper-poll/handlers/index.ts
@@ -1,2 +1,3 @@
 export * from './endedPollsHandler';
 export * from './interactionExecuteHandler';
+export * from './pollReactionHandler';
